backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
pymongo
python-multipart
python-jose
